added users and all_posts blueprints
