+++ admin/templates/admin/listposts.html
@@ -6,7 +6,7 @@
 <ul class="list-posts">
 {% for p in list_posts %}
 <li>
-    <p class="title"><a href="{{ url_for('show_post', alias=p.url)}}">{{p.title}}</a></p>
+    <p class="title"><a href="{{ url_for('all_posts.show_post', alias=p.url)}}">{{p.title}}</a></p>
     <p class="annonce">{{ p.text[:50] | striptags }}</p>
 </li>
 {% endfor %}
