+++ all_posts/templates/all_posts/confirm_delete.html
@@ -1,0 +1,5 @@
+{%block content%}
+<h1>Вы уверены, что хотите удалить этот пост?</h1>
+<p><a href="{{url_for('.delete_post', alias=alias)}}" method="delete">Да</a></p>
+<p><a href="{{url_for('.show_post', alias=alias)}}">Нет</a></p>
+{% endblock %}
