+++ all_posts/templates/all_posts/create_post.html
@@ -11,7 +11,7 @@
         {% endfor -%}
     </ul>
 {% endblock off_menu %}
-<form action="{{url_for('create_post')}}" method="post" class="form-contact">
+<form action="{{url_for('.create_post')}}" method="post" class="form-contact">
 <p><label>Название поста: </label><input type="text" name="name" value="" required />
 <p><label>Текст поста: </label>
 <p><textarea name="post" rows="7" cols="40"></textarea>
