+++ starter.py
@@ -1,15 +1,14 @@
-from flask import Flask, render_template, url_for, request, flash, session, redirect, abort, g, make_response
+from flask import Flask, render_template, request, flash, g
 import sqlite3
 import os
 from config import Config
 from FDataBase import FDataBase
 import datetime
-from werkzeug.security import generate_password_hash, check_password_hash
-from flask_login import LoginManager, login_required, login_user, logout_user, current_user
+from flask_login import LoginManager
 from UserLogin import UserLogin
-import base64
-from forms import AuthorisationForm, RegistrationForm
 from admin.admin import admin
+from all_posts.all_posts import all_posts
+from users.users import users
 
 
 app = Flask(__name__)
@@ -17,6 +16,9 @@ app.config.from_object(Config)
 app.config.update(dict(DATABASE=os.path.join(app.root_path,'gensh.db')))
 #app.permanent_session_lifetime = datetime.timedelta(days=10) #для запоминания сессии. Потом включить
 app.register_blueprint(admin, url_prefix='/admin')
+app.register_blueprint(all_posts)
+app.register_blueprint(users)
+
 login_manager = LoginManager(app)
 login_manager.login_view = 'authorisation'
 login_manager.login_message='Авторизуйтесь, чтобы просматривать эту страницу'
@@ -41,25 +43,6 @@ def get_db():
         g.link_db = connect_db()
         return g.link_db
 
-def get_avatars_dict(url, app=app):
-    username = dbase.getCommentatorsNames(url)
-    usernames = []
-    try:
-        for i in username:
-            usernames.append(i[0])
-        avas_dict = {}
-        for j, k in dbase.getCommentatorsAvas(usernames):
-            if k == None:
-                with app.open_resource(app.root_path + url_for('static', filename= 'images/default.jpeg'), "rb") as f:
-                    base64_string=base64.b64encode(f.read()).decode('utf-8')
-            else:
-                base64_string = base64.b64encode(k).decode('utf-8')
-            img_url=f'data:image/png;base64,{base64_string}'
-            avas_dict[j] = img_url
-    except:
-        pass
-    return avas_dict
-
 @app.teardown_appcontext
 def close_db(error):
     if hasattr(g, 'link_db'):
@@ -82,134 +65,6 @@ def load_user(user_id):
 @app.route("/")
 def index():
     return render_template("index.html", off_menu=dbase.getOffmenu())
-
-@app.route("/posts")
-def posts():
-    return render_template("posts.html",title = "Список постов", off_menu=dbase.getOffmenu(), posts=dbase.getPostsAnonce())
-
-@app.route("/post/<alias>", methods=['POST', 'GET'])
-def show_post(alias):
-    title, post, url, userid = dbase.get_post(alias)
-    if not title:
-        abort(404)
-    if request.method == "POST":
-        if len(request.form['comment'])> 1:
-            res = dbase.create_comment(request.form['comment'], alias)
-            if not res:
-                flash('Ошибка добавления комментария', category = 'error')
-            else:
-                return(redirect(url_for('show_post', alias=url)))
-        else:
-            flash('Ошибка добавления комментария', category = 'error')
-    return render_template("post.html", title = title, post=post, userid=str(userid), off_menu=dbase.getOffmenu(), comments=dbase.getCommentsAnonce(url), url=[url], avatars=get_avatars_dict(url))
-
-@app.route("/confirm_delete/<alias>")
-@login_required
-def confirm_delete(alias):
-    return render_template("confirm_delete.html", alias=alias)
-
-@app.route("/delete_post/<alias>", methods=['GET','DELETE'])
-@login_required
-def delete_post(alias):
-        dbase.delete_post(alias)
-        return(redirect(url_for('posts')))
-
-@app.route("/delete_comment/<alias>?<id>", methods=['GET', 'DELETE'])
-@login_required
-def delete_comment(alias, id):
-    dbase.delete_comment(id)
-    return(redirect(url_for('show_post', alias=alias)))
-
-@app.route("/my_posts")
-def my_posts():
-    return render_template("my_posts.html",title = "My Posts", off_menu=dbase.getOffmenu())
-
-@app.route("/create_post", methods=['POST', 'GET'])
-@login_required
-def create_post():
-    if request.method == "POST":
-        if len(request.form['name']) > 4 and len(request.form['post'])>1: #проверку на свой вкус
-            userid = int(current_user.get_id())
-            res = dbase.create_post(request.form['name'], request.form['post'], userid)
-            if not  res:
-                flash('Ошибка добавления статьи', category = 'error')
-            else:
-                return redirect(url_for('posts'))
-        else:
-            flash('Ошибка добавления статьи', category = 'error')
-    return render_template("create_post.html",title = "Create Post", off_menu=dbase.getOffmenu())
-
-@app.route('/logout')
-def logout():
-    logout_user()
-    return redirect(url_for('index'))
-
-
-@app.route("/profile/<username>")
-@login_required
-def profile(username):
-    if username != current_user.get_id():
-        abort(401)
-    return render_template("profile.html",title = "Profile")
-
-@app.route('/userava')
-@login_required
-def userava():
-    img=current_user.getAvatar(app)
-    if not img:
-        return ""
-    h=make_response(img)
-    h.headers['Content-Type'] = 'image/png'
-    return h
-
-@app.route('/upload', methods=["POST", "GET"])
-@login_required
-def upload():
-    if request.method == 'POST':
-        file = request.files['file']
-        if file and current_user.verifyExt(file.filename):
-            try:
-                img = file.read()
-                res = dbase.updateUserAvatar(img, current_user.get_id())
-                if not res:
-                    flash("Ошибка обновления аватара", "error")
-                flash("Аватар обновлен", "success")
-            except FileNotFoundError as e:
-                flash("Ошибка чтения файла", "error")
-        else:
-            flash("Ошибка обновления аватара", "error")
-    return redirect(url_for('profile', username=current_user.get_id()))
-
-@app.route("/authorisation", methods=['POST', 'GET'])
-def authorisation():
-    #session.permanent = True #для запоминания сессии. потом включить
-    if current_user.is_authenticated:
-        return redirect(request.args.get("next") or url_for("profile", username=current_user.get_id()))
-    form = AuthorisationForm()
-    if form.validate_on_submit():
-        user = dbase.getUserByLogin(form.name.data)
-        if user and check_password_hash(user['password'], form.password.data):
-            userlogin = UserLogin().create(user)
-            rm = form.remember.data
-            login_user(userlogin, remember=rm)
-            return redirect(request.args.get("next") or url_for("index"))
-        flash("Неверный логин или пароль", "error")
-    return render_template("authorisation.html", title="Authorisation", form=form) 
-
-@app.route("/register", methods=["POST", "GET"])
-def register():
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        hash = generate_password_hash(form.password.data)
-        res = dbase.add_user(form.name.data, hash, form.email.data)
-        if res:
-            flash("Вы успешно зарегистрированы!", "success")
-            return redirect(url_for('authorisation'))
-        else:
-            flash("Ошибка при добавлении в БД", "error")
-    else:
-        flash("Проверьте поле 'e-mail'. Также пароль должен быть не менее одного символа", "error")
-    return render_template("register.html", title="Registration", form=form)
 
 @app.route("/guides")
 def guides():
