+++ templates/base.html
@@ -20,14 +20,14 @@
         {% block mainmenu -%}
     <ul class="mainmenu">
     <li><a href="{{url_for('index')}}">Главная</a></li>
-    <li><a href="{{url_for('posts')}}">Посты</a></li>
-    <li><a href="{{url_for('create_post')}}">Создать Пост</a></li>
+    <li><a href="{{url_for('all_posts.posts')}}">Посты</a></li>
+    <li><a href="{{url_for('all_posts.create_post')}}">Создать Пост</a></li>
     {% if current_user.is_authenticated %}
-    <li><a href="{{url_for('profile', username=current_user.get_id())}}">Профиль</a></li>
-    <li><a href="{{url_for('logout')}}">Выйти</a></li>
+    <li><a href="{{url_for('users.profile', username=current_user.get_id())}}">Профиль</a></li>
+    <li><a href="{{url_for('users.logout')}}">Выйти</a></li>
     {% else -%}
-    <li><a href="{{url_for('authorisation')}}">Авторизоваться</a></li>
-    <li><a href="{{url_for('register')}}">Зарегистрироваться</a></li>
+    <li><a href="{{url_for('users.authorisation')}}">Авторизоваться</a></li>
+    <li><a href="{{url_for('users.register')}}">Зарегистрироваться</a></li>
     </ul>
         {% endif %}
         {% endblock mainmenu %}
