--- templates/confirm_delete.html
@@ -1,5 +1,0 @@
-{%block content%}
-<h1>Вы уверены, что хотите удалить этот пост?</h1>
-<p><a href="{{url_for('delete_post', alias=alias)}}" method="delete">Да</a></p>
-<p><a href="{{url_for('show_post', alias=alias)}}">Нет</a></p>
-{% endblock %}
