+++ users/templates/users/authorisation.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link type="text/css" href="{{ url_for('static', filename='css/styles.css')}}" rel="stylesheet" />
+    <link type="text/css" href="{{ url_for('.static', filename='css/styles.css')}}" rel="stylesheet" />
     {% if title %}
     <title>Genshin Funzone - {{ title }}</title>
     {% else %}
@@ -37,7 +37,7 @@
         <p>{{ form.remember.label() }} {{ form.remember() }}
         <p>{{ form.submit() }}
         <hr align=left width="300px">
-        <p><a href="{{url_for('register')}}">Регистрация</a>
+        <p><a href="{{url_for('.register')}}">Регистрация</a>
         </form>
     </div>
     {% endblock content %}
